Topology file for Classic L2 Acc/Agg/Core arch
This topology contains two pairs of acc-agg clusters connected by a 2
layer backbone. To maximize interop and test case scenarios for suzieq,
I made one of the clusters be eos, one cluster nxos, and the backbone
is made up for Cumulus and CSR. All of this can be made a single device
obviously by changing the device types.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>